requirements.txt
# SyftBox Privacy Filter App Dependencies

# Core framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Privacy and security
pyyaml>=6.0

# Machine learning for semantic analysis (Phase 2+)
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# LLM integration (Phase 4)
transformers>=4.30.0
torch>=2.0.0
ollama>=0.1.0

# HTTP and SSE
sse-starlette>=1.6.0
httpx>=0.25.0

# Utilities
python-multipart>=0.0.6
python-json-logger>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0